feature(react-scripts): include ui webpack configs (#2)
* Use ContextReplacementPlugin for moment locales in production

* v3.4.0

* Add initial server configs and fix malformed string in line 360

* v3.4.1

* Fix new keyword in ContextReplacementPlugin

* v3.4.2

* Fix filename and chunkFilename keys

* v3.4.3

* Fix typo in chunkFilename, common!! I can do this!!

* v3.4.4

* Add build path to server config

* v3.4.5

* Fix buildPath in build script

* v3.4.6

* Add server file to the template

* v1.0.6

* Add prettier as dependency to the template

* Fix helmet and change description in the template

* Add start-production-server to the scripts

* Add lengio prefix when pointing dependencies to local

* Disable eslint rules that might conflict with or prettier setup

* Move start-production-server and add prettier scripts

* Add prettierrc file to the template

* Add husky and lint-staged scripts to the template

* v3.4.7

* v1.0.7

* Move prettier to react-scripts instead of the template

* Move isInGitRepository function from init command to and util

* Create ensurePrettierGitHook function and use it on prettier command

* Add actual pre-commit file in ensurePrettierGitHook

* use ensurePrettierGitHook in init cmd

* v3.4.8

* Remove devDependencies from template (husky, lint-staged)

* Add uses strict to prettier cmd

* Move prettier rules from template to prettier script

* v3.4.9

* Add prettier cmd to package.json when initing a new app

* Add CopyWebpackPlugin to move static folder to the build

* Add simple server cmd

* v3.4.10

* Add missing dependencies

* v3.4.11

* v3.5.0

* v3.5.11

* v5.3.0

* v1.1.7

* v10.3.0

* Fix broken publicPath

* v3.5.12

* Add missisng prettier peer denpendency

* v5.3.1

* Update @lengio/eslint-config-react-app

* v3.5.13

* Add missing dependency for eslint-config package eslint-config-prettier

* v3.5.14

* Add missing peer dependency eslint-config-prettier

* v5.3.2

* Remove server script and file from template

* Add slang-atoms to paths

* Add slang-atoms package to resolve modules and apply webpack rules

* Add slang-atoms

* Add slang-atoms typescript aliases

* v1.1.8

* v3.5.15

* Remove devDependencies from blacklisted keys

* Add some missing dependencies and move some to devDependencies

* v3.5.16

* Upgrade slang-atoms package

* v3.5.17

* v1.1.9

* Upgrade slang-atoms version

* v1.1.10

* add devDependencies to templatePackageToMerge

* v1.1.11

* Ignore css typescript definitions

* Add serve script to init cmd

* v3.5.18

* v1.1.12

* Fix broken paths in serve

* v3.5.19

* Remove redundant path.join from the server

* Fix server path typo

* Remove CopyWebpackPlugin to move static forlder (will use public intead)

* v3.5.20

* Remove devDependencies from templatePackageToMerge (we want to replace)

* v3.5.21

* Move ssr webpack configs to its own file

* Use ssr webpack config in build only when the server file exists

* Improve serve script feedback

* add httpProxy constants to utils

* Add express server with some configs needed for ssr

* Move slang-stoms to peer denpendencies

* v3.5.22

* Improve prettier cmd (only read staged files)

* v3.5.23

* Move slang-atoms to dependencies

* v3.5.24

* Moved slang-atoms to peerDependencies (yep.. AGAIN)

* v3.5.25

* Fix prettier script when no staged files

* v3.5.26

* feature(slang-atoms): build slang-atoms separately  (#3)

+++ packages/react-scripts/scripts/utils/verifyTypeScriptSetup.js
@@ -131,7 +131,7 @@ function verifyTypeScriptSetup() {
     paths: { value: undefined, reason: 'aliased imports are not supported' },
     // Resolve src folder
     baseUrl: { suggested: './src' },
-    rootDirs: { suggested: ['./src'] }
+    rootDirs: { suggested: ['./src'] },
   };
 
   const formatDiagnosticHost = {
@@ -255,12 +255,14 @@ function verifyTypeScriptSetup() {
     writeJson(paths.appTsConfig, appTsConfig);
   }
 
-  // Reference `react-scripts` types
-  if (!fs.existsSync(paths.appTypeDeclarations)) {
-    fs.writeFileSync(
-      paths.appTypeDeclarations,
-      `/// <reference types="react-scripts" />${os.EOL}`
-    );
+  if (fs.existsSync(paths.appSrc)) {
+    // Reference `react-scripts` types
+    if (!fs.existsSync(paths.appTypeDeclarations)) {
+      fs.writeFileSync(
+        paths.appTypeDeclarations,
+        `/// <reference types="react-scripts" />${os.EOL}`
+      );
+    }
   }
 }
